Extract JWT utils
