+++ test/lizard-sign.test.ts
@@ -1,6 +1,8 @@
+/* eslint-disable @typescript-eslint/no-floating-promises */
 import {test} from 'node:test';
 import {lizardSign} from '../src/lizard-sign.js';
 import {assertSuccessfulResultFormat} from './assert-utils.js';
+import {looksLikeJwt} from './jwt-utils.js';
 
 const encodeSecret = (secret: string): Uint8Array =>
   new TextEncoder().encode(secret);
@@ -40,39 +42,17 @@ const lizardStrong = lizardSign('test', {
   strength: 'strong',
 });
 
-const looksLikeJwt = (prefix: string) => (value: string) => {
-  if (!value.startsWith(prefix)) {
-    return false;
-  }
-
-  const splitted = value.split('.');
-  const hasTwoDots = splitted.length === 3;
-  if (
-    !hasTwoDots ||
-    splitted[0] === undefined ||
-    splitted[1] === undefined ||
-    splitted[2] === undefined
-  ) {
-    return false;
-  }
-
-  if (splitted[0].length < 15) {
-    return false;
-  }
-
-  if (splitted[1].length < 30) {
-    return false;
-  }
-
-  if (splitted[2].length < 30) {
-    return false;
-  }
-
-  return true;
-};
-
 test('lizard should generate JWT 256', async () => {
   const actual = await lizard({id: 'lunar123'});
+  assertSuccessfulResultFormat(
+    actual,
+    looksLikeJwt('test:'),
+    'Should look like a JWT token'
+  );
+});
+
+test('lizard should generate JWT 256 with context', async () => {
+  const actual = await lizard({id: 'lunar123', team: 'red'});
   assertSuccessfulResultFormat(
     actual,
     looksLikeJwt('test:'),
